requirements.txt
Pillow==9.4.0
pillow-heif==0.10.0